+++ AgoraScene_iOS/AgoraScene_iOS/APP/en.lproj/Localizable.strings
@@ -107,4 +107,4 @@
 "Standard" = "Standard";
 "Spatial Audio" = "Spatial Audio";
 "Power natural conversations that make people feel like they're 'in the room' together" = "Power natural conversations that make people feel like they're 'in the room' together";
-
+"Current Customer Usage" = "Current Customer Usage";
