+++ AgoraScene_iOS/AgoraScene_iOS/APP/zh-Hans.lproj/Localizable.strings
@@ -116,3 +116,4 @@
 "If you have any query, please feel free to contact" = "若您有任何疑问，请联系";
 "Welcome to " = "欢迎来到";
 "Room Closed" = "房间销毁";
+"Current Customer Usage" = "以下客户都在用";
