+++ AgoraScene_iOS/AgoraScene_iOS/Business/VoiceChat2.0/Controller/VoiceRoomViewController+ChatBar.swift
@@ -1,0 +1,264 @@
+//
+//  VoiceRoomViewController+ChatBar.swift
+//  AgoraScene_iOS
+//
+//  Created by 朱继超 on 2022/10/11.
+//
+
+import Foundation
+import SVGAPlayer
+
+extension VoiceRoomViewController {
+    
+    func showEQView() {
+        preView = VMPresentView(frame: CGRect(x: 0, y: ScreenHeight, width: ScreenWidth, height: 450~))
+        preView.isAudience = !isOwner
+        preView.roomInfo = roomInfo
+        preView.ains_state = ains_state
+        preView.selBlock = {[weak self] state in
+            self?.ains_state = state
+            self?.rtckit.setAINS(with: state)
+            if self?.isOwner == false {return}
+            if let use_robot = self?.roomInfo?.room?.use_robot{
+                if use_robot == false {
+                    self?.view.makeToast("请房主先激活机器人")
+                    return
+                }
+                if state == .high {
+                    self?.rtckit.playMusic(with: .ainsHigh)
+                } else if state == .mid {
+                    self?.rtckit.playMusic(with: .ainsMid)
+                } else {
+                    self?.rtckit.playMusic(with: .ainsOff)
+                }
+            }
+        }
+        preView.useRobotBlock = {[weak self] flag in
+            if self?.alienCanPlay == true && flag == true {
+                self?.rtckit.playMusic(with: .alien)
+            }
+            
+            if self?.alienCanPlay == true && flag == false {
+                self?.rtckit.stopPlayMusic()
+            }
+            
+            self?.activeAlien(flag)
+        }
+        preView.volBlock = {[weak self] vol in
+            self?.updateVolume(vol)
+        }
+        preView.eqView.effectClickBlock = {[weak self] in
+            guard let effect = self?.roomInfo?.room?.sound_effect else {return}
+            self?.rtckit.playMusic(with: self?.getSceneType(effect) ?? .social)
+        }
+        preView.eqView.soundBlock = {[weak self] index in
+            if self?.isOwner == false {return}
+            if let use_robot = self?.roomInfo?.room?.use_robot{
+                if use_robot == false {
+                    self?.view.makeToast("请房主先激活机器人")
+                    return
+                }
+            }
+            let count = (index - 1000) / 10
+            let tag = (index - 1000) % 10
+            self?.rtckit.playSound(with: count, type: tag == 1 ? .ainsOff : .ainsHigh)
+            self?.rtcView.showAlienMicView = .blue
+        }
+        self.view.addSubview(preView)
+        self.isShowPreSentView = true
+        self.sRtcView.isUserInteractionEnabled = false
+        self.rtcView.isUserInteractionEnabled = false
+        self.headerView.isUserInteractionEnabled = false
+        
+        UIView.animate(withDuration: 0.5, animations: {
+            self.preView.frame = CGRect(x: 0, y: ScreenHeight - 450~, width: ScreenWidth, height: 450~)
+        }, completion: nil)
+    }
+    
+    func applyMembersAlert() {
+        let userAlert = VoiceRoomUserView(frame: CGRect(x: 0, y: 0, width: ScreenWidth, height: 420),controllers: [VoiceRoomApplyUsersViewController(roomId: self.roomInfo?.room?.room_id ?? ""),VoiceRoomInviteUsersController(roomId: self.roomInfo?.room?.room_id ?? "")],titles: [LanguageManager.localValue(key: "Raised Hands"),LanguageManager.localValue(key: "Invite On-Stage")]).cornerRadius(20, [.topLeft,.topRight], .white, 0)
+        let vc = VoiceRoomAlertViewController(compent: PresentedViewComponent(contentSize: CGSize(width: ScreenWidth, height: 420)), custom: userAlert)
+        self.presentViewController(vc)
+    }
+    
+    func showGiftAlert() {
+        let giftsAlert = VoiceRoomGiftsView(frame: CGRect(x: 0, y: 0, width: ScreenWidth, height: (110/84.0)*((ScreenWidth-30)/4.0)+180), gifts: self.gifts()).backgroundColor(.white).cornerRadius(20, [.topLeft,.topRight], .clear, 0)
+        let vc = VoiceRoomAlertViewController(compent: PresentedViewComponent(contentSize: CGSize(width: ScreenWidth, height: (110/84.0)*((ScreenWidth-30)/4.0)+180)), custom: giftsAlert)
+        giftsAlert.sendClosure = { [weak self] in
+            self?.sendGift(gift: $0)
+            if $0.gift_id == "VoiceRoomGift9" {
+                vc.dismiss(animated: true)
+                self?.rocketAnimation()
+            }
+        }
+        self.presentViewController(vc)
+    }
+    
+    func sendGift(gift: VoiceRoomGiftEntity) {
+        gift.userName = VoiceRoomUserInfo.shared.user?.name ?? ""
+        gift.portrait = VoiceRoomUserInfo.shared.user?.portrait ?? self.userAvatar
+        var giftList: VoiceRoomGiftView? = self.view.viewWithTag(1111) as? VoiceRoomGiftView
+        if giftList == nil {
+            giftList = self.giftList()
+            self.view.addSubview(giftList!)
+        }
+        giftList?.gifts.append(gift)
+        giftList?.cellAnimation()
+        if let chatroom_id = self.roomInfo?.room?.chatroom_id,let uid = self.roomInfo?.room?.owner?.uid,let id = gift.gift_id,let name = gift.gift_name,let value = gift.gift_price,let count = gift.gift_count {
+            VoiceRoomIMManager.shared?.sendCustomMessage(roomId: chatroom_id, event: VoiceRoomGift, customExt: ["gift_id":id,"gift_name":name,"gift_price":value,"gift_count":count,"userNaem":VoiceRoomUserInfo.shared.user?.name ?? "","portrait":VoiceRoomUserInfo.shared.user?.portrait ?? self.userAvatar], completion: { message, error in
+                if error == nil,message != nil {
+                    if let c = Int(count),let v = Int(value),var amount = VoiceRoomUserInfo.shared.user?.amount {
+                        amount += c*v
+                        VoiceRoomUserInfo.shared.user?.amount = amount
+                    }
+                    self.notifyServerGiftInfo(id: id, count: count, uid: uid)
+                } else {
+                    self.view.makeToast("Send failed \(error?.errorDescription ?? "")",point: self.toastPoint, title: nil, image: nil, completion: nil)
+                }
+            })
+        }
+    }
+    
+    func notifyServerGiftInfo(id: String,count: String,uid: String) {
+        if let roomId = self.roomInfo?.room?.room_id {
+            VoiceRoomBusinessRequest.shared.sendPOSTRequest(api: .giftTo(roomId: roomId), params: ["gift_id":id,"num":Int(count) ?? 1,"to_uid":uid]) { dic, error in
+                if let result = dic?["result"] as? Bool,error == nil,result {
+                    debugPrint("result:\(result)")
+                } else {
+                    self.view.makeToast("Send failed!",point: self.toastPoint, title: nil, image: nil, completion: nil)
+                }
+            }
+        }
+    }
+    
+    func rocketAnimation() {
+        let player = SVGAPlayer(frame: CGRect(x: 0, y: 0, width: ScreenWidth, height: ScreenHeight))
+        player.loops = 1
+        player.clearsAfterStop = true
+        player.contentMode = .scaleAspectFill
+        player.delegate = self
+        player.tag(199)
+        self.view.addSubview(player)
+        let parser = SVGAParser()
+        parser.parse(withNamed: "rocket", in: .main) { entitiy in
+            player.videoItem = entitiy
+            player.startAnimation()
+        } failureBlock: { error in
+            player.removeFromSuperview()
+        }
+    }
+    
+    func changeHandsUpState() {
+        if self.isOwner {
+            self.applyMembersAlert()
+            self.chatBar.refresh(event: .handsUp, state: .selected, asCreator: true)
+        } else {
+            if self.chatBar.handsState == .unSelected {
+                self.userApplyAlert(nil)
+            } else if self.chatBar.handsState == .selected {
+                self.userCancelApplyAlert()
+            }
+        }
+    }
+    
+    func changeMicState() {
+        self.chatBar.micState = !self.chatBar.micState
+        self.chatBar.refresh(event: .mic, state: self.chatBar.micState ? .selected:.unSelected, asCreator: false)
+        //需要根据麦位特殊处理
+        self.chatBar.micState == true ? self.muteLocal(with: 0):self.unmuteLocal(with: 0)
+    }
+    
+    func showUsers() {
+        let contributes = VoiceRoomUserView(frame: CGRect(x: 0, y: 0, width: ScreenWidth, height: 420),controllers: [VoiceRoomGiftersViewController(roomId: self.roomInfo?.room?.room_id ?? "")],titles: [LanguageManager.localValue(key: "Contribution List")]).cornerRadius(20, [.topLeft,.topRight], .white, 0)
+        let vc = VoiceRoomAlertViewController(compent: PresentedViewComponent(contentSize: CGSize(width: ScreenWidth, height: 420)), custom: contributes)
+        self.presentViewController(vc)
+    }
+    
+    func showApplyAlert(_ index: Int) {
+        let isHairScreen = SwiftyFitsize.isFullScreen
+        let manageView = VMManagerView(frame: CGRect(x: 0, y: 0, width: ScreenWidth, height:isHairScreen ? 264~ : 264~ - 34))
+        guard let mic_info = roomInfo?.mic_info?[index] else {return}
+        manageView.micInfo = mic_info
+        manageView.resBlock = {[weak self] (state, flag) in
+            self?.dismiss(animated: true)
+            if state == .invite {
+                if flag {
+                    self?.applyMembersAlert()
+                } else {
+                    self?.kickoff(with: index)
+                }
+            } else if state == .mute {
+                if flag {
+                    self?.mute(with: index)
+                } else {
+                    self?.unMute(with: index)
+                }
+            } else {
+                if flag {
+                    self?.lock(with: index)
+                } else {
+                    self?.unLock(with: index)
+                }
+            }
+        }
+        let vc = VoiceRoomAlertViewController.init(compent: PresentedViewComponent(contentSize: CGSize(width: ScreenWidth, height: isHairScreen ? 264~ : 264~ - 34)), custom: manageView)
+        self.presentViewController(vc)
+    }
+    
+    func userApplyAlert(_ index: Int?) {
+        let applyAlert = VoiceRoomApplyAlert(frame: CGRect(x: 0, y: 0, width: ScreenWidth, height: (205/375.0)*ScreenWidth),content: "Request to Speak?",cancel: "Cancel",confirm: "Confirm",position: .bottom).backgroundColor(.white).cornerRadius(20, [.topLeft,.topRight], .clear, 0)
+        let vc = VoiceRoomAlertViewController(compent: PresentedViewComponent(contentSize: CGSize(width: ScreenWidth, height: (205/375.0)*ScreenWidth)), custom: applyAlert)
+        applyAlert.actionEvents = { [weak self] in
+            if $0 == 31 {
+                self?.requestSpeak(index: index)
+            }
+            vc.dismiss(animated: true)
+        }
+        self.presentViewController(vc)
+    }
+    
+    func requestSpeak(index: Int?) {
+        guard let roomId = self.roomInfo?.room?.room_id else { return }
+        VoiceRoomBusinessRequest.shared.sendPOSTRequest(api: .submitApply(roomId: roomId), params: index != nil ? ["mic_index":index ?? 2]:[:]) { dic, error in
+            if error == nil,dic != nil,let result = dic?["result"] as? Bool {
+                if result {
+                    self.chatBar.refresh(event: .handsUp, state: .selected, asCreator: false)
+                    self.view.makeToast("Apply success!", point: self.toastPoint, title: nil, image: nil, completion: nil)
+                } else {
+                    self.view.makeToast("Apply failed!", point: self.toastPoint, title: nil, image: nil, completion: nil)
+                }
+            } else {
+                self.view.makeToast("\(error?.localizedDescription ?? "")", point: self.toastPoint, title: nil, image: nil, completion: nil)
+            }
+        }
+    }
+    
+    func cancelRequestSpeak(index: Int?) {
+        guard let roomId = self.roomInfo?.room?.room_id else { return }
+        VoiceRoomBusinessRequest.shared.sendDELETERequest(api: .cancelApply(roomId: roomId), params: [:]) { dic, error in
+            if error == nil,dic != nil,let result = dic?["result"] as? Bool {
+                if result {
+                    self.view.makeToast("Cancel Apply success!", point: self.toastPoint, title: nil, image: nil, completion: nil)
+                    self.chatBar.refresh(event: .handsUp, state: .unSelected, asCreator: false)
+                } else {
+                    self.view.makeToast("Cancel Apply failed!", point: self.toastPoint, title: nil, image: nil, completion: nil)
+                }
+            } else {
+                self.view.makeToast("\(error?.localizedDescription ?? "")",point: self.toastPoint, title: nil, image: nil, completion: nil)
+            }
+        }
+    }
+    
+    func userCancelApplyAlert() {
+        let cancelAlert = VoiceRoomCancelAlert(frame: CGRect(x: 0, y: 0, width: ScreenWidth, height: (205/375.0)*ScreenWidth)).backgroundColor(.white).cornerRadius(20, [.topLeft,.topRight], .clear, 0)
+        let vc = VoiceRoomAlertViewController(compent: PresentedViewComponent(contentSize: CGSize(width: ScreenWidth, height: (205/375.0)*ScreenWidth)), custom: cancelAlert)
+        cancelAlert.actionEvents = { [weak self] in
+            if $0 == 30 {
+                self?.cancelRequestSpeak(index: nil)
+            }
+            vc.dismiss(animated: true)
+        }
+        self.presentViewController(vc)
+    }
+    
+}
