+++ AgoraScene_iOS/AgoraScene_iOS/Common/Utils/VRVerifyCodeNumberView.swift
@@ -14,7 +14,7 @@ class VRVerifyCodeNumberView: UIView {
     var cursorColor = UIColor(0x009FFF)
     
     lazy var numLabel: UILabel = {
-        UILabel(frame: CGRect(x: 0, y: 0, width: self.frame.width, height: self.frame.height)).textAlignment(.center).font(.systemFont(ofSize: 18))
+        UILabel(frame: CGRect(x: 0, y: 0, width: self.frame.width, height: self.frame.height)).textAlignment(.center).font(.systemFont(ofSize: 18)).textColor(.darkText)
     }()
     
     /// 光标
